+++ python/mujoco/__init__.py
@@ -124,38 +124,60 @@ def from_zip(file: Union[str, IO[bytes]]) -> _specs.MjSpec:
   if not zipfile.is_zipfile(file):
     raise ValueError(f'File {file} is not a zip file.')
   with zipfile.ZipFile(file, 'r') as zip_file:
+    xml_dir = None
     for zip_info in zip_file.infolist():
       if not zip_info.filename.endswith(os.path.sep):
         with zip_file.open(zip_info.filename) as f:
           if zip_info.filename.endswith('.xml'):
             xml_string = f.read()
+            xml_dir = os.path.dirname(zip_info.filename)
           else:
             assets[zip_info.filename] = f.read()
+
   if not xml_string:
     raise ValueError('No XML file found in zip file.')
+
+  relative_assets = {}
+  for key, value in assets.items():
+    new_key = os.path.relpath(key, xml_dir)
+    relative_assets[new_key] = value
+  assets = relative_assets
+
   return _specs.MjSpec.from_string(xml_string, assets=assets)
 
 
 class _MjBindModel:
+  """Wrapper for MjModel that allows binding multiple specs."""
+
   def __init__(self, elements: Sequence[Any]):
-    self.elements = elements
+    object.__setattr__(self, 'elements', elements)
 
   def __getattr__(self, key: str):
     items = []
     for e in self.elements:
       items.extend(getattr(e, key))
     return items
+
+  def __setattr__(self, key: str, value: Any):
+    raise AttributeError(f'Cannot set {key} on MjModel.')
 
 
 class _MjBindData:
+  """Wrapper for MjData that allows binding multiple specs."""
+
   def __init__(self, elements: Sequence[Any]):
-    self.elements = elements
+    object.__setattr__(self, 'elements', elements)
 
   def __getattr__(self, key: str):
     items = []
     for e in self.elements:
       items.extend(getattr(e, key))
     return items
+
+  def __setattr__(self, key: str, value: Any):
+    value_it = iter(value)
+    for element in self.elements:
+      setattr(element, key, next(value_it))
 
 
 def _bind_model(